+++ rpi_ai/main.py
@@ -11,10 +11,10 @@ from flask import Flask, Response, jsonify, request
 from waitress import serve
 from werkzeug.datastructures import FileStorage, Headers, ImmutableMultiDict
 
-from rpi_ai.api_types import Message, SpeechResponse
 from rpi_ai.chatbot import Chatbot
 from rpi_ai.config import ChatbotConfig, Config
 from rpi_ai.functions import FUNCTIONS
+from rpi_ai.models import Message, SpeechResponse
 
 logging.basicConfig(
     format="[%(asctime)s] %(levelname)s in %(module)s: %(message)s", datefmt="%d/%m/%Y | %H:%M:%S", level=logging.INFO
